+++ front/src/views/manager/graph/Graph.js
@@ -6,7 +6,7 @@ import PageContainer from '../../../components/container/PageContainer';
 import TodayJoinMembers from './components/TodayJoinMembers';
 import MonthJoinMembers from './components/MonthJoinMembers';
 import YearJoinMembers from './components/YearJoinMembers';
-import MembersLine from './components/MembersLine';
+import MemberAgeDonut from './components/MemberAgeDonut';
 import SubscribeMembersLine from './components/SubscribeMembersLine';
 import CountSomething from './components/CountSomething'
 import mainContext from "../main/contexts/MainContext";
@@ -30,9 +30,8 @@ const Dashboard = () => {
           <Grid item xs={12} lg={4}>
             <YearJoinMembers />
           </Grid>
-
           <Grid item xs={12} lg={6}>
-            <MembersLine />
+            <MemberAgeDonut />
           </Grid>
           <Grid item xs={12} lg={6}>
             <SubscribeMembersLine />
